create utils for combining fields and fieldGroups to reuse in models
