+++ ui/app/models/pki-certificate-sign.js
@@ -2,7 +2,7 @@ import { copy } from 'ember-copy';
 import { computed } from '@ember/object';
 import DS from 'ember-data';
 import Certificate from './pki-certificate';
-
+import { combineFieldGroups } from 'vault/utils/openapi-to-attrs';
 const { attr } = DS;
 
 export default Certificate.extend({
@@ -18,11 +18,14 @@ export default Certificate.extend({
 
   fieldGroups: computed('signVerbatim', function() {
     const options = { Options: ['altNames', 'ipSans', 'ttl', 'excludeCnFromSans', 'otherSans'] };
-    const groups = [
+    let groups = [
       {
         default: ['csr', 'commonName', 'format', 'signVerbatim'],
       },
     ];
+    if (this.newFields.length) {
+      groups = combineFieldGroups(groups, this.newFields, []);
+    }
     if (this.get('signVerbatim') === false) {
       groups.push(options);
     }
