+++ ui/app/models/pki-certificate.js
@@ -3,6 +3,7 @@ import { computed } from '@ember/object';
 import DS from 'ember-data';
 import lazyCapabilities, { apiPath } from 'vault/macros/lazy-capabilities';
 import fieldToAttrs, { expandAttributeMeta } from 'vault/utils/field-to-attrs';
+import { combineFieldGroups } from 'vault/utils/openapi-to-attrs';
 
 const { attr } = DS;
 
@@ -30,73 +31,22 @@ export default DS.Model.extend({
   role: attr('object', {
     readOnly: true,
   }),
-
-  // revocationTime: attr('number'),
-  // commonName: attr('string', {
-  //   label: 'Common Name',
-  // }),
-
-  // altNames: attr('string', {
-  //   label: 'DNS/Email Subject Alternative Names (SANs)',
-  // }),
-
-  // ipSans: attr('string', {
-  //   label: 'IP Subject Alternative Names (SANs)',
-  // }),
   otherSans: attr({
-    // editType: 'stringArray',
-    // label: 'Other SANs',
     helpText:
       'The format is the same as OpenSSL: <oid>;<type>:<value> where the only current valid type is UTF8',
   }),
-
-  // ttl: attr({
-  //   label: 'TTL',
-  //   editType: 'ttl',
-  // }),
-
-  // format: attr('string', {
-  //   defaultValue: 'pem',
-  //   possibleValues: ['pem', 'der', 'pem_bundle'],
-  // }),
-
-  // excludeCnFromSans: attr('boolean', {
-  //   label: 'Exclude Common Name from Subject Alternative Names (SANs)',
-  //   defaultValue: false,
-  // }),
-
-  // certificate: attr('string'),
-  // issuingCa: attr('string', {
-  //   label: 'Issuing CA',
-  // }),
-  // caChain: attr('string', {
-  //   label: 'CA chain',
-  // }),
-  // privateKey: attr('string'),
-  // privateKeyType: attr('string'),
-  // serialNumber: attr('string'),
 
   fieldsToAttrs(fieldGroups) {
     return fieldToAttrs(this, fieldGroups);
   },
 
   fieldDefinition: computed('newFields', function() {
-    const groups = [
+    let groups = [
       { default: ['commonName', 'format'] },
       { Options: ['altNames', 'ipSans', 'ttl', 'excludeCnFromSans', 'otherSans'] },
     ];
-    if (this.newFields) {
-      let allFields = [];
-      for (let group in groups) {
-        let type = Object.keys(groups[group])[0];
-        allFields.concat(groups[group]);
-      }
-      let otherFields = this.newFields.filter(field => {
-        !allFields.includes(field);
-      });
-      if (otherFields.length) {
-        groups.default.concat(otherFields);
-      }
+    if (this.newFields.length) {
+      groups = combineFieldGroups(groups, this.newFields, []);
     }
     return groups;
   }),
